requirements.txt
# Infraestrutura e API
uvicorn==0.41.0
fastapi==0.135.1
pydantic==2.12.5

# LLM (Groq)
groq==1.1.0

# Inteligência Artificial e Embeddings (Local)
torch==2.10.0
transformers==5.3.0
sentence-transformers==5.2.3
accelerate==1.13.0

# LangChain e Integrações
langchain==1.2.10
langchain-community==0.4.1
langchain-core==1.2.17
langchain-huggingface==1.2.1
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.1
langgraph==1.0.10

# Banco de Dados Vetorial
qdrant-client==1.17.0

# Processamento de Documentos
PyPDF2==3.0.1
python-docx==1.2.0
python-pptx==1.0.2
tiktoken==0.12.0

# Interface Web
streamlit==1.55.0

# Outras Dependências
python-dotenv==1.0.1

